back/requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
drf-nested-routers==0.92.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
packaging==20.4
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
